+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed an issue where we would emit stdout after a failure, even when the stdout
   had been emitted as part of the task
 - Added `outFile` and `enableExperimentalMirrors` to the dart2js task
+- Removed the deprecated `defaultInit` method.
 
 ## 0.7.2 (2015/7/8)
 - Auto-expand directories passed to `Analyzer.analyze()` to a list of Dart